Renamed k8s-related import for clarity and consistency
Examples:
- `k8s.io/api/apps/v1` => `appsAPI`
- `k8s.io/api/core/v1` => `coreAPI`
